Mark unreachable points in map and friends (map)
The reimplementation of map in private/map.rkt calls to the primitive
map to let it generate the errors. The optimizer doesn't understand
that this will generate the error, so we add an explicit error in the
unreachable points. This allows the type information generated in the
map implementation to be propagated to the later code.