optimizer: Use some functions from #%kernel in the tests
Some primitives from #%kernel are shadowed by racket/base. Some tests need
the primitive to make the compilation more predictable. Other tests are
about properties of the primitives tha can't be tested in the other version.

Most of this tests we already here, but the procedures are renamed to make
the dependency more explicit.