fix reduction of constructors in a boolean context
The expressions like (not (vector (x) (y))) are  reduced to
(begin (x) (y) #f) to avoid the creation of the unnecessary vector.

Fix the test of the predicate that recognize if the type of the constructed
object implies #t or #f. Apparently in any case this created an error, but it's
necessary to fix this to add boolean? to the list of relevant predicates.